requirements.txt
PyQt5>=5.15.0
pandas>=1.3.0
openpyxl>=3.0.0
psycopg2-binary>=2.9.0 